refactor(vscode): dedupe parseBounds + diff-mode bar across webview
Two follow-ups flagged in earlier PRs (#774 / the migration audit) — both
were "the same helper exists in two places, fix when convenient".

## parseBounds

Was duplicated between `cardData.ts` (exported, 11 lines) and a private
copy in `a11yOverlay.ts` (same body). The cardData copy is the
canonical one; `a11yOverlay.ts` now imports from it and drops the
local re-implementation.

`a11yOverlay.ts`: 223 → 205 lines (−18).

## buildDiffModeBar

Was duplicated between `preview/diffOverlay.ts` (`buildDiffModeBar`) and
`history/behavior.ts` (`buildHistoryDiffModeBar`). Same DOM shape, same
CSS class (`.diff-mode-bar`), same `aria-selected` toggle pattern, same
3-tab Side / Overlay / Onion list — only the persistence wiring
differed and that already lives outside the function via the
`onChange` callback. The only meaningful semantic difference was that
the preview copy used a typed `for` loop and the history copy used
`forEach` for the click handler's selectAll walk; standardised on the
`for` form.

Lifts both into `shared/diffModeBar.ts` (62 lines) exporting
`buildDiffModeBar(initialMode, onChange) → HTMLElement` plus the
`DiffMode = "side" | "overlay" | "onion"` union. `preview/diffOverlay.ts`
re-exports `DiffMode` from the shared module so existing imports
(`import { DiffMode } from "./diffOverlay"`) keep working without a
ripple-edit.

The stack / pane builders (`buildDiffStack` / `buildPreviewDiffPane` /
`buildHistoryDiffStack` / `buildDiffPane`) intentionally *not*
deduped — they use different CSS class prefixes
(`preview-diff-pane` vs `diff-pane`) so the two panels' diff overlays
can be styled independently, and there's no shared CSS truth to
collapse them around.

`preview/diffOverlay.ts`: 378 → 341 lines (−37).
`history/behavior.ts`: 841 → 802 lines (−39).

Net: −32 LOC across the webview tree (94 saved minus 62 new).

## Test plan

- [x] `npm run compile` clean (host + webview)
- [x] `npm test` — 326/326 passing
- [x] `npm run format:check` clean
- [ ] Manual smoke:
  - Live panel: open a focused diff (Diff vs HEAD / vs main), cycle
    Side → Overlay → Onion tabs, verify persisted-mode survives a
    panel reload.
  - History panel: expand a row's "Diff vs previous" / "Diff vs current"
    and cycle the same three tabs, verify persisted-mode survives a
    panel reload.
  - Verify the a11y overlay still paints — `parseBounds` fires for
    each finding/node and the overlay layer renders boxes at
    percentage coordinates.
